Changes: 1. double->float; 2. association if's; 3. add more tickRate and changes tickRate; 4. https://github.com/Glitchfiend/BiomesOPlenty/commit/7cb684b541aff97d66ee68b1bf371c936491e138 5. switch -> if else if and commeted out switch code; 6. reduce getBlock()/getBlockMetadata()/getBlockLightOpacity(); 7. revert part old commit(s); 8. https://pmd.github.io/pmd-6.1.0/pmd_rules_java_design.html#simplifiedternary 9. return old sacred oak plank;

+++ src/main/java/biomesoplenty/api/biome/BOPBiomeDecorator.java
@@ -88,19 +88,19 @@ public class BOPBiomeDecorator<T extends BiomeFeatures> extends BiomeDecorator
     
     public static <T extends WorldGenerator> T getRandomWeightedWorldGenerator(HashMap<T, ? extends Number> worldGeneratorMap)
     {
-        double completeWeight = 0D;
+        float completeWeight = 0F;
 
         for (Number weight : worldGeneratorMap.values())
         {
-            completeWeight += Double.parseDouble(String.valueOf(weight));
+            completeWeight += Float.parseFloat(String.valueOf(weight));
         }
 
-        double random = (new XSTR()).nextFloat() * completeWeight;
-        double countWeight = 0D;
+        float random = (new XSTR()).nextFloat() * completeWeight;
+        float countWeight = 0F;
 
         for (Map.Entry<T, ? extends Number> entry : worldGeneratorMap.entrySet())
         {
-            countWeight += Double.parseDouble(String.valueOf(entry.getValue()));
+            countWeight += Float.parseFloat(String.valueOf(entry.getValue()));
 
             if (countWeight >= random) { return entry.getKey(); }
         }
